Split up nodes
