+++ packages/shadenfreude/src/nodes/index.ts
@@ -1,135 +1,9 @@
-import { Color } from "three"
-import { float, node, nodeFactory, variable, vec3 } from "../factories"
-import { Operator, Program, Value, Variable } from "../types"
-
-export const TimeNode = nodeFactory(() => {
-  const u_time = variable("float", 0)
-
-  return node({
-    name: "Time Uniform",
-    uniforms: { u_time },
-    outputs: { value: float(u_time) },
-
-    update: (_, dt) => {
-      ;(u_time.value as number)! += dt
-    }
-  })
-})
-
-export const ColorNode = nodeFactory<{ color?: Value<"vec3"> }>(
-  ({ color = new Color(1.0, 1.0, 1.0) }) => ({
-    name: "Constant Color Value",
-    outputs: {
-      value: vec3(color)
-    }
-  })
-)
-
-export const FloatNode = nodeFactory<{ value: Value<"float"> }>(
-  ({ value }) => ({
-    name: "Constant Float Value",
-    outputs: {
-      value: float(value)
-    }
-  })
-)
-
-export const PositionNode = nodeFactory(() => ({
-  name: "Vertex Position",
-  outputs: { value: vec3() },
-  vertex: {
-    header: "varying vec3 v_position;",
-    body: "value = v_position = position;"
-  },
-  fragment: {
-    header: "varying vec3 v_position;",
-    body: "value = v_position;"
-  }
-}))
-
-export const CSMMasterNode = nodeFactory<{
-  position?: Value<"vec3">
-  normal?: Value<"vec3">
-  pointSize?: Value<"float">
-  diffuseColor?: Value<"vec3">
-  fragColor?: Value<"vec3">
-  emissiveColor?: Value<"vec3">
-}>(({ position, normal, pointSize, diffuseColor, fragColor, emissiveColor }) =>
-  node({
-    name: "Master Node",
-    inputs: {
-      diffuseColor: vec3(diffuseColor),
-      emissiveColor: vec3(emissiveColor),
-      position: vec3(position),
-      normal: vec3(normal)
-    },
-    vertex: {
-      body: `
-        ${pointSize ? "csm_PointSize = pointSize;" : ""}
-        ${position ? "csm_Position = position;" : ""}
-        ${normal ? "csm_Normal = normal;" : ""}
-      `
-    },
-    fragment: {
-      body: `
-        ${diffuseColor ? "csm_DiffuseColor.rgb = diffuseColor;" : ""}
-        ${fragColor ? "csm_FragColor.rgb = fragColor;" : ""}
-        ${emissiveColor ? "csm_EmissiveColor.rgb = emissiveColor;" : ""}
-      `
-    }
-  })
-)
-
-export const OperatorNode = (
-  operator: Operator,
-  inputs: { a: Variable; b: Variable }
-) =>
-  node({
-    name: `Perform ${operator} on ${inputs.a.type}`,
-    inputs: {
-      a: variable(inputs.a.type, inputs.a),
-      b: variable(inputs.b.type, inputs.b)
-    },
-    outputs: {
-      value: variable(inputs.a.type)
-    },
-    vertex: {
-      body: `value = a ${operator} b;`
-    },
-    fragment: {
-      body: `value = a ${operator} b;`
-    }
-  })
-
-export const AddNode = nodeFactory<{ a: Variable; b: Variable }>(({ a, b }) =>
-  OperatorNode("+", { a, b })
-)
-
-export const MultiplyNode = nodeFactory<{ a: Variable; b: Variable }>(
-  ({ a, b }) => OperatorNode("*", { a, b })
-)
-
-export const MixNode = nodeFactory<{
-  a: Variable
-  b: Variable
-  factor: Value<"float">
-}>(({ a, b, factor }) => ({
-  name: "Mix",
-  inputs: {
-    a: variable(a.type, a),
-    b: variable(b.type, b),
-    factor: variable("float", factor)
-  },
-  outputs: {
-    value: variable(a.type)
-  },
-  vertex: {
-    body: `value = mix(a, b, factor);`
-  },
-  fragment: {
-    body: `value = mix(a, b, factor);`
-  }
-}))
+import { float, nodeFactory, vec3 } from "../factories"
+import { Program, Value } from "../types"
+export * from "./inputs"
+export * from "./masters"
+export * from "./math"
+export * from "./values"
 
 export const FresnelNode = nodeFactory(() => ({
   name: "Fresnel",
